Move deprecated adapted digikikify to a different file
