+++ packages/voictents-and-estinants-engine/src/custom/programs/find-unused-exports/markUnusedExports.ts
@@ -262,6 +262,12 @@ export const markUnusedExports = buildEstinant({
             'packages/voictents-and-estinants-engine/src/custom/programs/render-knowledge-graph/file/getAssociatedFileGraphElements.ts',
           importedIdentifierName: 'getAssociatedFileGraphElements',
         },
+        {
+          importingFilePath: '',
+          importedFilePath:
+            'packages/voictents-and-estinants-engine/src/type-script-adapter/deprecatedDigikikify.ts',
+          importedIdentifierName: 'digikikify',
+        },
       ];
 
     [
